+++ services/src/main/java/org/keycloak/protocol/oidc/AccessTokenIntrospectionProvider.java
@@ -18,6 +18,7 @@
 package org.keycloak.protocol.oidc;
 
 import java.nio.charset.StandardCharsets;
+import java.util.Arrays;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Map;
@@ -31,6 +32,7 @@ import java.util.Base64;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.io.IOUtils;
+import org.keycloak.OAuth2Constants;
 import org.keycloak.broker.oidc.OIDCIdentityProvider;
 import org.keycloak.broker.oidc.OIDCIdentityProviderConfig;
 import org.keycloak.broker.provider.util.SimpleHttp;
@@ -43,6 +45,7 @@ import org.keycloak.common.util.Time;
 import org.keycloak.connections.httpclient.HttpClientProvider;
 import org.keycloak.crypto.SignatureProvider;
 import org.keycloak.crypto.SignatureVerifierContext;
+import org.keycloak.models.ClientScopeModel;
 import org.keycloak.models.IdentityProviderModel;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.ProtocolMapperModel;
@@ -51,6 +54,7 @@ import org.keycloak.models.RealmModel;
 import org.keycloak.models.UserModel;
 import org.keycloak.models.customcache.CustomCacheProvider;
 import org.keycloak.models.customcache.CustomCacheProviderFactory;
+import org.keycloak.models.utils.KeycloakModelUtils;
 import org.keycloak.protocol.oidc.representations.OIDCConfigurationRepresentation;
 import org.keycloak.models.UserSessionModel;
 import org.keycloak.protocol.ProtocolMapper;
@@ -85,7 +89,7 @@ public class AccessTokenIntrospectionProvider implements TokenIntrospectionProvi
         this.session = session;
         this.realm = session.getContext().getRealm();
         this.tokenManager = new TokenManager();
-        initTokenCache();
+       // initTokenCache();
     }
 
     private void initTokenCache(){
@@ -107,7 +111,10 @@ public class AccessTokenIntrospectionProvider implements TokenIntrospectionProvi
             if (realmUrl.equals(issuer)) {
                 return introspectKeycloak(token);
             } else {
-                if (isExpired(tokenJson.get("exp").asLong())) {
+                long exp = tokenJson.get("exp").asLong();
+                if (isExpired(exp)) {
+                    String clientId = tokenJson.get("azp").asText();
+                    logger.warn("Token introspection for "+clientId+" client has expired  token. Token expiration = "+exp+ ". Current time = "+ Time.currentTime());
                     ObjectNode tokenMetadata = JsonSerialization.createObjectNode();
                     tokenMetadata.put("active", false);
                     return Response.ok(JsonSerialization.writeValueAsBytes(tokenMetadata)).type(MediaType.APPLICATION_JSON_TYPE).build();
@@ -128,9 +135,10 @@ public class AccessTokenIntrospectionProvider implements TokenIntrospectionProvi
     }
 
     protected Response introspectKeycloak (String token) {
+        AccessToken accessToken = null;
         try {
 
-            AccessToken accessToken = verifyAccessToken(token);
+            accessToken = verifyAccessToken(token);
             ObjectNode tokenMetadata;
 
             if (accessToken != null) {
@@ -149,13 +157,14 @@ public class AccessTokenIntrospectionProvider implements TokenIntrospectionProvi
                 }
             } else {
                 tokenMetadata = JsonSerialization.createObjectNode();
+                logger.warn("Keycloak token introspection return null access token.");
             }
 
             tokenMetadata.put("active", accessToken != null);
-
             return Response.ok(JsonSerialization.writeValueAsBytes(tokenMetadata)).type(MediaType.APPLICATION_JSON_TYPE).build();
         } catch (Exception e) {
-            logger.warn("Exception during Keycloak introspection",e);
+            String clientId = accessToken != null ? accessToken.getIssuedFor() : "unknown";
+            logger.warn("Exception during Keycloak introspection for "+clientId+" client.",e);
             throw new RuntimeException("Error creating token introspection response.", e);
         }
     }
@@ -186,6 +195,10 @@ public class AccessTokenIntrospectionProvider implements TokenIntrospectionProvi
         //client - user exist - otherwise validation failed before
         ClientModel client = realm.getClientByClientId(token.getIssuedFor());
         UserModel user = getUserFromToken(token, client);
+        if (user == null ) {
+            logger.warn("Token introspection enhancement failed for "+token.getIssuedFor()+" client: User can not retrieve from access token");
+            return null;
+        }
         ClientContextUtils ccu= new ClientContextUtils(client,token.getScope(),session,user);
 
         AtomicReference<AccessToken> finalToken = new AtomicReference<>(token);
@@ -209,7 +222,10 @@ public class AccessTokenIntrospectionProvider implements TokenIntrospectionProvi
         UserModel user = tokenManager.lookupUserFromStatelessToken(session, realm, token);
         //if can not get user from token, search user from token sessionState
         if (user == null) {
-            user = new UserSessionCrossDCManager(session).getUserSessionWithClient(realm, token.getSessionState(), false, client.getId()).getUser();
+            boolean offlineTokenRequested =  (token.getScope() != null && Arrays.stream(token.getScope().split(" ")).anyMatch(x -> OAuth2Constants.OFFLINE_ACCESS.equals(x))) || client.getClientScopes(true).values().stream().anyMatch(x -> OAuth2Constants.OFFLINE_ACCESS.equals(x.getName()));
+            UserSessionModel userSession = new UserSessionCrossDCManager(session).getUserSessionWithClient(realm, token.getSessionState(), offlineTokenRequested, client.getId());
+            if (userSession != null)
+                user = userSession.getUser();
         }
         return user;
     }
@@ -217,9 +233,9 @@ public class AccessTokenIntrospectionProvider implements TokenIntrospectionProvi
     protected Response introspectWithExternal(String token, String issuer, RealmModel realm) throws IOException {
 
         try {
-            String cachedToken = (String) tokenRelayCache.get(new Key(token, realm.getName()));
-            if (cachedToken != null)
-                return Response.ok(cachedToken).type(MediaType.APPLICATION_JSON_TYPE).build();
+//            String cachedToken = (String) tokenRelayCache.get(new Key(token, realm.getName()));
+//            if (cachedToken != null)
+//                return Response.ok(cachedToken).type(MediaType.APPLICATION_JSON_TYPE).build();
 
             IdentityProviderModel issuerIdp = realm.getIdentityProvidersStream().filter(idp -> issuer.equals(idp.getConfig().get("issuer"))).findAny().orElse(null);
             if (issuerIdp != null) {
@@ -244,7 +260,7 @@ public class AccessTokenIntrospectionProvider implements TokenIntrospectionProvi
             //if failed to find issuer in IdPs or IntrospectionEndpoint does not exist for specific Idp return false
             logger.warn(issuerIdp != null ? "Remote introspection: problem getting remote Idp with issuer " + issuer + "introspection endpoint" : "Remote introspection: Idp with issuer " + issuer + " does not exist");
             ObjectNode tokenMetadata = JsonSerialization.createObjectNode();
-            tokenMetadata.put("active", false);
+           // tokenMetadata.put("active", false);
             return Response.ok(JsonSerialization.writeValueAsBytes(tokenMetadata)).type(MediaType.APPLICATION_JSON_TYPE).build();
         } catch (Exception e) {
             logger.warn("Error during remote introspection", e);